[Gold V] Title: 토마토, Time: 84 ms, Memory: 6112 KB -BaekjoonHub

+++ 백준/Gold/7576. 토마토/토마토.cc
@@ -4,7 +4,7 @@
 using namespace std;
 
 int n, m;
-queue<pair<int, int>> que;
+queue<pair<short, short>> que;
 char arr[1000][1000];
 bool visited[1000][1000];
 int clear; //익은 토마토 개수
